# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/survey_survey_restrict_users:
    get:
      tags:
        - SurveySurveyRestrictUsers
      summary: List SurveySurveyRestrictUsers
      description: Get a paginated list of SurveySurveyRestrictUsers with optional filtering and sorting
      operationId: listSurveySurveyRestrictUsers
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SurveySurveyRestrictUserList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - SurveySurveyRestrictUsers
      summary: Create SurveySurveyRestrictUser
      description: Create a new SurveySurveyRestrictUser
      operationId: createSurveySurveyRestrictUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSurveySurveyRestrictUserRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SurveySurveyRestrictUser'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/survey_survey_langs:
    get:
      tags:
        - SurveySurveyLangs
      summary: List SurveySurveyLangs
      description: Get a paginated list of SurveySurveyLangs with optional filtering and sorting
      operationId: listSurveySurveyLangs
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SurveySurveyLangList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - SurveySurveyLangs
      summary: Create SurveySurveyLang
      description: Create a new SurveySurveyLang
      operationId: createSurveySurveyLang
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSurveySurveyLangRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SurveySurveyLang'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    SurveySurveyRestrictUser:
      type: object
      required:
        - id
        - survey_id
        - user_id
        - metadata
      properties:
        id:
          type: string
          format: uuid
        survey_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        metadata:
          type: object
...
    CreateSurveySurveyRestrictUserRequest:
      type: object
      required:
        - survey_id
        - user_id
        - metadata
      properties:
        survey_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        metadata:
          type: object
...
    SurveySurveyRestrictUserList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/SurveySurveyRestrictUser'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    SurveySurveyLang:
      type: object
      required:
        - id
        - survey_id
        - lang_id
        - metadata
      properties:
        id:
          type: string
          format: uuid
        survey_id:
          type: string
          format: uuid
        lang_id:
          type: string
          format: uuid
        metadata:
          type: object
...
    CreateSurveySurveyLangRequest:
      type: object
      required:
        - survey_id
        - lang_id
        - metadata
      properties:
        survey_id:
          type: string
          format: uuid
        lang_id:
          type: string
          format: uuid
        metadata:
          type: object
...
    SurveySurveyLangList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/SurveySurveyLang'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []